Move to a Fedora 30 based container image
Given the issues that have arisen from the Debian OpenJDK packages (and
subsequent Docker builds), the build image has been switched to a Fedora
container with a Red Hat OpenJDK build. The reasoning here is the Java
packaging team at Debian does not have a great track record when it comes to
producing reliable packages.

To summarize:

  * Debian pushed buggy pre-release packages to stable and stable-backports.
  * Debian was not involved in producing the OpenJDK Docker images at
    DockerHub.
  * The OpenJDK image published to DockerHub incorrectly uses the
    OpenJDK trademark, when it is not an official product of the OpenJDK project.