Use Java 11 instead of Java 12 in the build image
There are more breaking changes (technically not really, as they were
internal APIs, but people are consuming them nevertheless) in Java 12 in
the form of APIs that have been stripped out.

Until the ecosystem is in a position where everyone can upgrade their
dependencies to the latest rolling release of Java without worrying
about compatability we should stick with whatever the latest LTS is.