Split out procedures to repeatable migrations
Moves each procedure and function definition into a separate file that
contains only its definition, making it easy to edit a function on the
fly.

The SQL files the procedures are stored in use the Flyway repeatable
migration naming scheme and will be re-run every time the database comes
up.