Explicitly baseline the database on startup
Re-enables baselining of the database on startup, to allow connecting to
and migrating against a database with a target schema that already
exists.

The reason this init listener exists is so baselining can be configured,
in the abscene of a `quarkus.flyway.baselineOnMigrate` option. If
baselining is not configured and the application is started with a
non-empty database then Flyway will refuse to run the migrations.

Baselining on migration should only be disabled in cases where you don't
want Flyway to create the `flyway_history` or run any of your migrations
in a database that has a non-empty schema (i.e. to prevent accidental
dataloss).